MealMate-replit/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "recipe-scrapers>=15.7.1",
    "requests>=2.32.3",
    "flask-login>=0.6.3",
    "oauthlib>=3.2.2",
    "pyjwt>=2.10.1",
    "werkzeug>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
    "sqlalchemy>=2.0.41",
    "psycopg2-binary>=2.9.10",
    "google-generativeai>=0.8.5",
    "trafilatura>=2.0.0",
    "gunicorn>=23.0.0",
]
